+++ services/foodservice/src/main/java/com/shop/foodservice/repo/FoodRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.shop.foodservice.model.Food;
 
-public interface FoodRepository extends JpaRepository<Food, Integer>{
+public interface FoodRepository extends JpaRepository<Food, Long>{
 
 }
